pin scipy version to avoid binary conflict for libstdc++.so.6
Python packages' binaries are linked to the libstdc++ shipped with Julia when loaded. If these binaries were built with newer versions of libstdc++, errors might arise. This should fix this for now, but a permanent solution that ensures that python packages use the binaries aof their corresponding conda environment would be better.